+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Static methods to ApiQuery class
+- Tests
 ### Changed
 ### Fixed
+- Validator's $valid response
 
 ## [0.9.2] - 2023-10-08
 ### Added